moved test files to own folder, and updated package json veriations
