+++ package.json
@@ -57,14 +57,14 @@
     "eslint": "^8.57.1",
     "eslint-config-react-app": "^7.0.1",
     "gh-pages": "^6.1.1",
-    "jsdom": "^27.4.0",
+    "jsdom": "^24.1.3",
     "vite": "^7.3.0",
     "vitest": "^4.0.16"
   },
   "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
   "main": "server.js",
   "engines": {
-    "node": "20.x",
+    "node": ">=20.19.0",
     "npm": "10.x"
   },
   "author": "",
