+++ src/test/HomePage.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import HomePage from './HomePage';
+import HomePage from '../pages/HomePage';
 
 describe('HomePage Component', () => {
   test('renders without crashing', () => {
